pyproject.toml
[project]
name = "fantasy-hockey"
version = "0.1.0"
description = "ESPN Fantasy Hockey data access"
requires-python = ">=3.13"
dependencies = [
    "espn-api>=0.45.1",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
]

[project.scripts]
fantasy = "fantasy_hockey.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/fantasy_hockey"]
